+++ website/docs/intro.md
@@ -11,7 +11,7 @@ Did you write or find an article that helped you get started with `Azure SQL`? O
 
 ### Option 1 - Ask
 
-Just open an issue mentioning the link you would like to add, a title and a short description.
+Just open an [issue](https://github.com/yorek/awesome-azure-sql/issues/new/choose) mentioning the link you would like to add, a title and a short description.
 
 ### Option 2 - Do
 
@@ -20,12 +20,13 @@ We welcome contributions of [requested resources](https://github.com/yorek/aweso
 #### Submit a resource
 
 1. Add an entry to [website/items/](https://github.com/yorek/awesome-azure-sql/blob/main/website/items/) *.yml* files that includes:
-    - **title** - A short title that reflects what the resource is about.
+    - **title:** - A short title that reflects what the resource is about.
     - **description:** - 1-2 sentence description of the resource.
-    - **url** - Link to the resource.
+    - **url:** - Link to the resource.
     - **icon:** - The icon is an emoji. The supported emojis are defined here: [Markdown emojis](https://github.com/markdown-templates/markdown-emojis).
-    - **author** - Author's name to credit on the gallery card
-    - **website**  - At least one tag is mandatory. You can add as many as you want. (I suggest not to add to many. Make sure the tags used a really helpful to identify the content. Try to limit the number of tags to 2 at maximum).
+    - **author:** - Author's name to credit on the gallery card.
+    - **website**: - Link to Author's GitHub or other relevant website used for attribution.
+    - **tags:**  - At least one tag is mandatory. You can add as many as you want. (I suggest not to add to many. Make sure the tags used a really helpful to identify the content. Try to limit the number of tags to 2 at maximum).
 2. Open a PR!
 
 #### Tags
@@ -36,5 +37,5 @@ If you want to add a nested tag, just separate the content using a slash. For ex
 
 Other than these, we always welcome feedback through a:
 
-- [**Feature request or idea**](https://github.com/yorek/awesome-azure-sql/issues): This may be a new resource or improvement to the Awesome Azure SQL.
-- [**bug report**](https://github.com/yorek/awesome-azure-sql/issues): Let us know if something is broken.
+- [**Feature request or idea**](https://github.com/yorek/awesome-azure-sql/issues/new/choose): This may be a new resource or improvement to the Awesome Azure SQL.
+- [**Bug report**](https://github.com/yorek/awesome-azure-sql/issues/new/choose): Let us know if something is broken.
